Spruce up Razor source generator
