+++ src/RazorSdk/Targets/Microsoft.NET.Sdk.Razor.SourceGenerators.targets
@@ -66,7 +66,7 @@ Copyright (c) .NET Foundation. All rights reserved.
       <CompilerVisibleProperty Include="GenerateRazorMetadataSourceChecksumAttributes" />
       <CompilerVisibleProperty Include="MSBuildProjectDirectory" />
       <CompilerVisibleProperty Include="_RazorSourceGeneratorDebug" />
-      <CompilerVisibleProperty Include="_RazorSourceGeneratorLog" />
+      <CompilerVisibleProperty Include="_RazorSourceGeneratorHeapDumps" />
     </ItemGroup>
 
     <ItemGroup>
